+++ Changelog.md
@@ -5,6 +5,7 @@
 - Change: The Cage Trap is now reusable
 - Fix: Cannot access the recipe book for the Password-protected Furnace
 - Fix: Password-protected Furnace is not shown as a recipe catalyst for furnace recipes in JEI
+- Fix: Changing the owner of the Inventory Scanner or doors does not change the owner of the other block/half
 - Misc.: More texture changes
 
 --------------------------Changelog for v1.9.0.2 of SecurityCraft--------------------------